% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{add_text}
\alias{add_text}
\title{Add text area to Moodle page}
\usage{
add_text(tab, section, title, text)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{section}{Scalar character vector giving the name of the section where the file should be added}

\item{title}{Scalar character vector giving the title displayed for the file on the main page}

\item{text}{A character vector holding text string you want to upload. Can contain raw HTML. If lenth > 1, the elements are concatenated together using a line break.}
}
\description{
Add text area to Moodle page
}
