% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mdl_cache.R
\name{check_mdl_cache}
\alias{check_mdl_cache}
\title{Check Cached Data}
\usage{
check_mdl_cache()
}
\value{
invisible(NULL)
}
\description{
Checks that the cached data is available.
}
\examples{
\dontrun{
check_mdl_cache()
}
}
