% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{create_groups}
\alias{create_groups}
\title{Create Moodle groups}
\usage{
create_groups(tab, groups)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{groups}{A character vector holding the names of each group to be created}
}
\description{
Create Moodle groups
}
