% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{create_new_section}
\alias{create_new_section}
\title{Create a new section on a Moodle page}
\usage{
create_new_section(tab, section_name)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{section_name}{Scalar character vector giving the name of the new section}
}
\value{
An (invisible) \link[httr:response]{httr::response} object containing the response from Moodle's core_courseformat_update_course endpoint
}
\description{
Create a new section on a Moodle page
}
