% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{dates}
\alias{dates}
\title{Populate a calendar listing}
\usage{
dates(
  begin = "2019-09-05",
  end = "2019-12-12",
  days_of_week = c(3, 5),
  holidays = NULL
)
}
\arguments{
\item{begin}{first day of class}

\item{end}{last day of class}

\item{days_of_week}{on which days does the class meet?}

\item{holidays}{vector of exceptions}
}
\description{
Populate a calendar listing
}
\examples{
dates()
dates(days_of_week = c(2, 4, 6),
      holidays = c("2019-10-14", "2019-10-15"))
}
