% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{download_quiz_responses}
\alias{download_quiz_responses}
\title{Download quiz responses and attachments}
\usage{
download_quiz_responses(
  tab,
  item_id,
  include_attachments = FALSE,
  output_dir,
  overwrite = FALSE
)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{item_id}{The ID number of the Moodle assignment/quiz. Can be found in the URL shown in the
address bar when the assignment/quiz is opened in the web browser.}

\item{include_attachments}{Logical scalar indicating whether attachments should be downloaded}

\item{output_dir}{Path to the folder where responses/attachments retrieved from Moodle will be
output. A folder matching the name of the quiz/assignment will be created in this folder, and
all responses attachments will be placed within. Has no effect when \code{include_attachments} is
\code{FALSE}.}

\item{overwrite}{Logical scalar. Controls whether existing files are overwritten when downloading
attachments}

\item{questions}{Vector of question numbers that are expected to have attachments}
}
\value{
A data frame beginning with the columns \code{Name}, \verb{Email address}, \code{Attempt}, \code{Status}, and
\code{Grade/100.00}, followed by a column for each quiz question response. The columns are named
\verb{Response 1}, \verb{Response 2}, \verb{Response 3}, etc.

Note that if a quiz allows a student multiple attempts, then all attempts will be included in
this data frame.
}
\description{
Downloads all complete responses to a Moodle quiz, along with any file attachments.
}
\details{
Text responses for each finished attempted are stored in the responses.csv file Within
\code{output_dir} a folder is created for each student, where the name of the folder corresponds to
the student's email address. Each student's file attachments from the Moodle quiz are placed
in the their respective folder.
}
