% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_config_file.R
\name{example_config_yml}
\alias{example_config_yml}
\title{Example Config yml File}
\usage{
example_config_yml()
}
\value{
invisible(TRUE) on success
}
\description{
Opens an example config file in the active editor.
If you already have a config.yml file cut-and-paste the "moodleR" section,
and update the respective fields.
Alternatively use "save as" to save this file as config.yml
It is recommended that you check the config with config::get()
}
\examples{
\dontrun{
example_config_yml()
}
}
