% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{export_gradebook}
\alias{export_gradebook}
\title{Retrieve course roster from Moodle}
\usage{
export_gradebook(tab, grade_items)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{grade_items}{Character vector holding the precise names of items in the Moodle gradebook; these names can be obtained from \code{\link[=get_gradebook_items]{get_gradebook_items()}}. When this argument is omitted, no items are retrieved and only the ID variables (name, email, etc.) are retrieved.}
}
\value{
A data frame beginning with 4 columns (\code{first_name}, \code{last_name}, \code{id_number}, and \code{email_address}), and additional columns for each requested item.
}
\description{
Retrieve the course roster from Moodle. This roster includes all students, but not all course
participants (i.e., does not include TA's, instructors, administrators, etc.).
}
\examples{
\dontrun{
tab <- open_moodle(
  site_url = "https://moodle.smith.edu",
  graphical = TRUE
)
# Export the full gradebook
grade_items <- get_gradebook_items(tab)
grades <- export_gradebook(tab, names(grade_items))

# Just the items starting with prefix "Reading Quiz"
is_quiz_item <- startsWith(names(grade_items), "Reading Quiz")
grades <- export_gradebook(tab2, names(grade_items)[is_quiz_item])
}
}
