% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{extract_cookies}
\alias{extract_cookies}
\title{Extract cookies from a Chrome browser tab}
\usage{
extract_cookies(tab)
}
\arguments{
\item{tab}{A \code{\link[chromote:ChromoteSession]{chromote::ChromoteSession()}} object}
}
\value{
an \link[httr:httr-package]{httr::httr} requests object with all fields NULL except the options$cookie field
}
\description{
Extract cookies from a Chrome browser tab
}
