% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_forum.R
\name{extract_edges.mdl_forum_data}
\alias{extract_edges.mdl_forum_data}
\title{Create an edgelist}
\usage{
\method{extract_edges}{mdl_forum_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_forum_data"}}

\item{...}{Further arguments passed on to methods}
}
\description{
Create an edgelist.
}
