% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{extract_edges}
\alias{extract_edges}
\title{Create an edgelist}
\usage{
extract_edges(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_forum_data"}}

\item{...}{Further arguments passed on to methods}

\item{prefix}{Defaults to \code{"mdl_"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}
}
\description{
Create an edgelist.
}
