% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{extract_items.mdl_quiz_data}
\alias{extract_items.mdl_quiz_data}
\title{Extract item-level and distractor data}
\usage{
\method{extract_items}{mdl_quiz_data}(x, marks = "categorical", fill = NA,
  mat = TRUE)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz_data"}}

\item{marks}{Char, defaults to "categorical"}

\item{fill}{What to use to fill missing values, defaults to \code{NA}}

\item{mat}{If \code{TRUE} (default), the result will be a \code{matrix} otherwise a \code{data.frame}}
}
\description{
Extract a matrix or data.frame of question marks
}
\details{
Extract item marks either as binary (correct/incorrect) or categorical data (for multiplechoice questions). If the test contains a mixture of binary and categorical items (such as truefalse and multichoice), than \code{marks = "binary"} will collapse multiple choice items into correct/incorrect (1/0). When using \code{marks = "categorical"}, mutliplechoice options will be preserved but a key needs to be obtained (see for details). Raw percentages as they appear in Moodle can also be extracted by \code{marks = "moodle"}.
}
