% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{extract_key.mdl_quiz_data}
\alias{extract_key.mdl_quiz_data}
\title{Extract key}
\usage{
\method{extract_key}{mdl_quiz_data}(x, question.type = NULL,
  complete = FALSE)
}
\arguments{
\item{x}{Object of class \code{"mdl_quiz_data"}}

\item{question.type}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}

\item{complete}{Should question name, text etc. be provided? Returns a \code{data.frame}}
}
\description{
Extract key
}
