% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{extract_key}
\alias{extract_key}
\title{Extract key}
\usage{
extract_key(x, distractors = NULL)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz_data"}}

\item{distractors}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}
}
\description{
Extract key (only when distractor data are available).
}
