% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{extract_moodle_context_string}
\alias{extract_moodle_context_string}
\title{Extract the Moodle 'context' string}
\usage{
extract_moodle_context_string(html)
}
\arguments{
\item{html}{An HTML document object, such as the kind returned by \code{\link[rvest:read_html]{rvest::read_html()}}}
}
\value{
A scalar character vector representing the 'context' string
}
\description{
Extract the Moodle 'context' string
}
