% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{extract_moodle_session_key}
\alias{extract_moodle_session_key}
\title{Extract the Moodle 'sesskey' string}
\usage{
extract_moodle_session_key(tab)
}
\arguments{
\item{tab}{A \code{\link[chromote:ChromoteSession]{chromote::ChromoteSession()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A scalar character vector representing the Moodle session key
}
\description{
Whenever you log into Moodle, Moodle sets a "session key" that is expected to be included with POST and GET requests to the server. Given a Chromote session object, This function retrieves the current session key from the browser's JavaScript environment.
}
