% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{extract_nodes}
\alias{extract_nodes}
\title{Create a nodelist}
\usage{
extract_nodes(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_forum_data"}}

\item{...}{Further arguments passed on to methods}
}
\description{
Create a nodelist.
}
