% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{feedback}
\alias{feedback}
\title{Generate feedback report for Moodle}
\usage{
feedback(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Generate feedback report for Moodle
}
\examples{
\dontrun{
key <- "1S0vfV_xkebyTkC2bmhOv-MDNMLulb95GFwfvI9ZPtGw"
if (gogglesheets) {
  grades <- key \%>\%
    gs_key() \%>\%
    gs_read()
}
feedback(grades)
}
}
