% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{generate_rcode}
\alias{generate_rcode}
\title{Generate r code text given a template.}
\usage{
generate_rcode(
  df.fct,
  cols = NULL,
  name = "df",
  type = c("tibble", "data.frame")
)
}
\arguments{
\item{df.fct}{Function which generates a single dataset that could be given
to a student. This function should not take any parameters.}

\item{cols}{columns in the result of \code{df.fct} that should not be
considered wildcards. Takes any form of \code{dplyr::tidyselect}.}

\item{name}{the name of the resulting dataset (default = \code{"df"}).}

\item{type}{one of \code{"tibble"} or \code{"data.frame"} that specifies the
type of output that should be created from the code.}
}
\value{
invisible
}
\description{
Wildcards hold the various problems in Moodle. When you want each student
to have a unique dataset, there are many wildcards to generate. You must
then specify to students how to read in this data, the code for which will
contain Moodle's notation for wildcards. This provides a way of writing such
code.  It is to be used alongside \code{\link{generate_wildcards}}.
}
\details{
This prints code to the screen using \code{cat}.
}
