% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genquiz.R
\name{genquiz}
\alias{genquiz}
\title{genquiz}
\usage{
genquiz(B = 1, fun, folder, problem = 0, Show = FALSE, ...)
}
\arguments{
\item{B}{how many quizzes?}

\item{fun}{name of the R routine that makes a quiz}

\item{folder}{where is the .R located?}

\item{problem}{(otional) which problem should be done?}

\item{Show}{(otional)want to see what it looks like?}
}
\description{
This function generates moodle quizzes
}
\examples{
genquiz(10, example1)
}
\keyword{moodle}
