% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course.R
\name{get_course_modules}
\alias{get_course_modules}
\title{Get Course Modules}
\usage{
get_course_modules(conn, course.id, module.names = TRUE, module.type = NULL,
  prefix = "mdl_")
}
\arguments{
\item{conn}{DB connection}

\item{course.id}{Vector of course IDs}

\item{module.names}{Should module names be also fetched (slower)? Defaults to \code{TRUE}}

\item{prefix}{Database table prefix, defaults to \code{"mdl_"}}
}
\description{
Get a list of modules within given course(s).
}
