% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{get_course_roster}
\alias{get_course_roster}
\title{Retrieve course roster from Moodle}
\usage{
get_course_roster(tab)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A data frame with 4 columns (\code{first_name}, \code{last_name}, \code{id_number}, and \code{email_address}).
Note that the ID number corresponds to the "school ID" number, not their internal Moodle "secret"
ID number.
}
\description{
Retrieve the course roster from Moodle. This roster includes all students, but not all course
participants (i.e., does not include TA's, instructors, administrators, etc.).
}
