% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course.R
\name{get_courses}
\alias{get_courses}
\title{Get Courses}
\usage{
get_courses(conn, prefix = "mdl_")
}
\arguments{
\item{conn}{DB connection}

\item{prefix}{Database table prefix, defaults to \code{"mdl_"}}
}
\description{
Get course names, ids, basic info.
}
