% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forum.R
\name{get_forum}
\alias{get_forum}
\title{Get Forum}
\usage{
get_forum(conn, forum.id, prefix = "mdl_", suppress.warnings = TRUE)
}
\arguments{
\item{conn}{DB connection}

\item{forum.id}{Forum ID (also in browser address bar)}

\item{prefix}{Database table prefix, defaults to \code{"mdl_"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}
}
\description{
Fetches basic forum object.
}
