% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{get_gradebook_items}
\alias{get_gradebook_items}
\title{Get gradebook items}
\usage{
get_gradebook_items(tab)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A named character vector with one element per item in the course gradebook. The elements of the vector correspond to the internal item ID numbers used in the Moodle database. The names of each element correspond to the user-facing names assigned to each item.
}
\description{
Retrieve the names and internal ID numbers for all items in the gradebook.
}
