% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{get_group_ids}
\alias{get_group_ids}
\title{Retrieve Moodle group names and corresponding IDs}
\usage{
get_group_ids(tab)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A data.frame object with the following variables:
\describe{
\item{group_idnumber}{The groups's internal Moodle ID number}
\item{group_name}{The group's user-facing name in Moodle}
}
}
\description{
Retrieve name and corresponding internal ID number for each group in a Moodle course.
Note that this function does retrieve the names of members in each group.
}
