% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_forum.R
\name{get_module_data.mdl_forum}
\alias{get_module_data.mdl_forum}
\title{Get Forum Data}
\usage{
\method{get_module_data}{mdl_forum}(x, prefix = "mdl_",
  suppress.warnings = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_forum"}}

\item{prefix}{Defaults to \code{"mdl_"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}

\item{...}{Further arguments passed on to methods}
}
\description{
Fetch posts and related forum data.
}
