% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{get_module_data.mdl_quiz}
\alias{get_module_data.mdl_quiz}
\title{Get Quiz Data}
\usage{
\method{get_module_data}{mdl_quiz}(x, attempt = "first",
  question.type = NULL, prefix = "mdl_", ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz"}}

\item{attempt}{Defaults to \code{"first"} but \code{"all"} and \code{"last"} are also allowed}

\item{question.type}{Char vector, if \code{NULL}, all items regardless of type will be fetched}

\item{prefix}{Defaults to \code{"mdl_"}}

\item{...}{Further arguments passed on to methods}
}
\description{
Fetch item data from quiz.
}
