% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{get_module_data}
\alias{get_module_data}
\title{Get module data}
\usage{
get_module_data(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_"}, e.g. \code{"mdl_quiz"}}

\item{...}{Further arguments passed on to methods}
}
\description{
Fetch data for a specific module.
}
