% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{get_question_key}
\alias{get_question_key}
\title{Get Question Key}
\usage{
get_question_key(conn, question.type, quiz.id, prefix = "mdl_",
  suppress.warnings = TRUE)
}
\arguments{
\item{conn}{Connection object}

\item{question.type}{Char vector}

\item{quiz.id}{Numeric vector}

\item{prefix}{Defaults to \code{"mdl_quiz"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}
}
\description{
Get question key.
}
