% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quiz.R
\name{get_quiz}
\alias{get_quiz}
\title{Get Quiz}
\usage{
get_quiz(conn, quiz.id, attempt.state = "finished", prefix = "mdl_",
  suppress.warnings = TRUE)
}
\arguments{
\item{conn}{DB connection}

\item{quiz.id}{Quiz ID (also in browser address bar)}

\item{attempt.state}{Defaults to \code{"finished"}}

\item{prefix}{Database table prefix, defaults to \code{"mdl_"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}
}
\description{
Fetches basic quiz object.
}
