% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_sa.R
\name{get_shortanswer}
\alias{get_shortanswer}
\title{Get shortanswer item data}
\usage{
get_shortanswer(conn, quiz.id, attempt.id, prefix = "mdl_",
  suppress.warnings = TRUE)
}
\arguments{
\item{conn}{Connection object}

\item{quiz.id}{Numeric vector of length 1}

\item{attempt.id}{Vector of attempt ids}

\item{prefix}{Defaults to \code{"mdl_quiz"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}
}
\description{
Get shertanswer item data.
}
