% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{get_student_secret_ids}
\alias{get_student_secret_ids}
\title{Retrieve Moodle student IDs}
\usage{
get_student_secret_ids(tab)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A data.frame object with the following variables:
\describe{
\item{student_id}{The student's internal Moodle ID number (i.e., their "secret" ID)}
\item{student_email}{The student's email address}
}
}
\description{
Retrieve the internal ID number for each student in a Moodle course.

Note that the Moodle "secret ID" is \emph{not} the same as the student ID number you might see in the gradebook when you download their grades, or on their ID card, etc.
}
