% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_sa.R
\name{get_truefalse}
\alias{get_truefalse}
\title{Get truefalse item data}
\usage{
get_truefalse(conn, quiz.id, attempt.id, prefix = "mdl_")
}
\arguments{
\item{conn}{Connection object}

\item{quiz.id}{Numeric vector of length 1}

\item{attempt.id}{Vector of attempt ids}

\item{prefix}{Defaults to \code{"mdl_quiz"}}
}
\description{
Get truefalse item data.
}
