% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{get_user_agent}
\alias{get_user_agent}
\title{Retrieve the browser's user agent string}
\usage{
get_user_agent(tab)
}
\arguments{
\item{tab}{A \code{\link[chromote:ChromoteSession]{chromote::ChromoteSession()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}
}
\value{
A scalar character vector holding the browser's user agent string
}
\description{
Retrieve the browser's user agent string
}
