% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_moodle.R
\name{html_to_moodle}
\alias{html_to_moodle}
\title{Convert HTML output into Moodle XML.}
\usage{
html_to_moodle(file)
}
\arguments{
\item{file}{the HTML file that should be converted.}
}
\value{
invisible
}
\description{
HTML code generated from knitting together an RMarkdown document which
corresponds to several questions is converted to Moodle XML. This function
writes an XML file with the same name as the input file.
}
