% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imoodleR.R
\name{imoodleR}
\alias{imoodleR}
\title{imoodleR}
\usage{
imoodleR(txt, w = 100, caps = TRUE, pts = 1)
}
\description{
This function runs the imoodleR shiny app
}
\examples{
shinymoodleR()
}
\keyword{CLOZE}
\keyword{moodle,}
