% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{install_makefile}
\alias{install_makefile}
\alias{summarize_assignment}
\alias{winnow_group_submissions}
\title{Install makefiles}
\usage{
install_makefile(path)

summarize_assignment(path)

winnow_group_submissions(path, regexp = "Group_[A-Z]+")
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{regexp}{Regular expression to match against folder names}
}
\description{
Install makefiles
}
\examples{
assignment_dir <- "~/Dropbox/git/sds192/student_info/s19/SDS192- 01_201903-Mini-project #1-394031"
if (fs::dir_exists(assignment_dir)) {
  install_makefile(assignment_dir)
  summarize_assignment(assignment_dir)
}
}
