% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{is.mdl_}
\alias{is.mdl_}
\alias{is.mdl_quiz}
\alias{is.mdl_quiz_data}
\title{Test mdl_ objects}
\usage{
is.mdl_quiz(x)

is.mdl_quiz_data(x)
}
\arguments{
\item{x}{Object to be tested}
}
\description{
These objects contain summary information about particular module instance.
}
