% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{join_key_ans}
\alias{join_key_ans}
\title{Join answers and keys}
\usage{
join_key_ans(key, ans)
}
\arguments{
\item{key}{Data from get_[question type]_key query}

\item{ans}{Data from get_[question type]_ans query}
}
\description{
This should work for: tf, sa. Questions that are missing will be replaced by
NA records.
}
