% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{mc}
\usage{
mc(options, w, which.true, pts = 1)
}
\arguments{
\item{options}{vector of choices}

\item{w}{vector of weights}

\item{which.true}{either which of the options gets 100%, 
or a logical value TRUE=first option, False=second option}

\item{pts}{how many points is question worth?}
}
\description{
This function generates the code for a multiple choice CLOZE question
}
\examples{
mc(c("Yes", "No"), c(100, 0), 10)
mc(3, which.true=(pval<0.05))
}
\keyword{moodle}
