% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_config.R
\name{mdl_config}
\alias{mdl_config}
\title{Moodle Configuration Table}
\usage{
mdl_config(con = mdl_get_connection())
}
\arguments{
\item{con}{database connection object}
}
\value{
A tibble
}
\description{
Get the complete moodle configuration table.
Different from most mdl_* functions this function calls dplyr::collect()
before returning (as a config-table is not going to cause memory overflow).
}
