% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_create_cache.R
\name{mdl_create_cache}
\alias{mdl_create_cache}
\title{Creates a Local Cache}
\usage{
mdl_create_cache(
  con = mdl_get_connection(use_cache = FALSE),
  format = c("sqlite"),
  tbl_prefix = "mdl_",
  output_dir = mdl_get_cache_dir(),
  output_filename = mdl_get_cache_filename()
)
}
\arguments{
\item{con}{a database connection (to a moodle database)}

\item{format}{output format (mysql, csv)}

\item{tbl_prefix}{Moodle DB table prefix}

\item{output_dir}{where does the cache go}

\item{output_filename}{filename (in the case of sqlite output), or prefix (in_case of CSV download)}
}
\description{
Create local cache from a  Moodle Database
}
