% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum_posts.R
\name{mdl_forum_posts}
\alias{mdl_forum_posts}
\title{Get Forum Posts}
\usage{
mdl_forum_posts(con = mdl_get_connection(), tbl_prefix = "mdl_")
}
\arguments{
\item{con}{a database connection object}

\item{tbl_prefix}{table prefix}
}
\value{
A dbplyr reference object.
}
\description{
Returns a reference to the (cached) forum_posts table, with the most relevant columns selected.
}
\details{
For convenience two additional columns: forum_name and thread_name; are added.
}
