% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_get_connection.R
\name{mdl_get_cache_connection}
\alias{mdl_get_cache_connection}
\title{Connection to Cached Moodle Data}
\usage{
mdl_get_cache_connection(access = c("RO", "RWC"))
}
\arguments{
\item{access}{Specifies RO or RWC accesss}
}
\value{
a DBI connection object
}
\description{
Connection to Cached Moodle Data
}
