% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl-cache-paths.R
\name{mdl_get_cache_dir}
\alias{mdl_get_cache_dir}
\title{Get the Cache Directory}
\usage{
mdl_get_cache_dir()
}
\value{
Character vector with path
}
\description{
The cache directory can be set in the config-file (moodleR:->cache_dir:), which is recommended.
If it is not present, or no config is found, a tempdir() will be returned.
}
