% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl-cache-paths.R
\name{mdl_get_cache_filename}
\alias{mdl_get_cache_filename}
\title{Get the Cache Filename}
\usage{
mdl_get_cache_filename()
}
\value{
Character vector with path
}
\description{
The cache filename ("mdl_cache.sqlite" by default), can be set in the config-file((moodleR:->cache_dir:).
}
