% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_get_connection.R
\name{mdl_get_connection}
\alias{mdl_get_connection}
\title{Connect to Moodle Data}
\usage{
mdl_get_connection(use_cache = TRUE, config = "default")
}
\arguments{
\item{use_cache}{If TRUE (the default) connection to the local cache is returned.}

\item{config}{Select configuration from config file}
}
\value{
a DBI connection object
}
\description{
Returns a connection to a Moodle database or the cached version if available.e
}
