% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_get_grades.R
\name{mdl_grades_query}
\alias{mdl_grades_query}
\title{Raw query for grades}
\usage{
mdl_grades_query(tbl_prefix)
}
\arguments{
\item{tbl_prefix}{for moodle}
}
\value{
character vector with moodle query
}
\description{
Raw query for grades
}
\keyword{internal}
