% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_log.R
\name{mdl_log}
\alias{mdl_log}
\title{Access Moodle's Logs}
\usage{
mdl_log(con = mdl_get_connection())
}
\arguments{
\item{con}{A database connection object}
}
\value{
A dbplyr reference object.
}
\description{
Returns a reference to the log-table.
}
