% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_get_posts.R
\name{mdl_posts_query}
\alias{mdl_posts_query}
\title{Raw query for discussion posts}
\usage{
mdl_posts_query(tbl_prefix)
}
\arguments{
\item{tbl_prefix}{for moodle}
}
\value{
character vector with moodle query
}
\description{
Raw query for discussion posts
}
\keyword{internal}
