% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_users.R
\name{mdl_users}
\alias{mdl_users}
\title{Get Moodle Users}
\usage{
mdl_users(con = mdl_get_connection())
}
\arguments{
\item{con}{a database connection object}
}
\value{
A dbplyr reference object
}
\description{
Returns a reference to the (cached) users table, with the most relevant columns selected.
}
