% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_to_word.R
\name{moodle_to_word}
\alias{moodle_to_word}
\title{Convert a Moodle quiz to a Microsoft Word paper survey.}
\usage{
moodle_to_word(file, solution = FALSE, title = "Quiz")
}
\arguments{
\item{file}{the Moodle XML file that should be converted.}

\item{solution}{if `TRUE`, solutions are included in the output. If `FALSE`
(default), solutions are not included.}

\item{title}{string indicating the title of the quiz (default = 'Quiz').}
}
\value{
invisible
}
\description{
Take a Moodle XML file and create a Microsoft Word version suitable for
distributing a paper quiz.
}
