% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_page.R
\name{MoodlePage}
\alias{MoodlePage}
\title{MoodlePage class}
\description{
MoodlePage objects represents a Chrome browser tab tone that is logged in to a Moodle instance.
A \emph{session} in a Chromote object. Note that in the Chrome
DevTools Protocol a session is a debugging session connected to a \emph{target},
which is a browser window/tab or an iframe.
}
\section{Super class}{
\code{\link[chromote:ChromoteSession]{chromote::ChromoteSession}} -> \code{MoodlePage}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{site_url}}{root URL for the Moodle instance your are connected to}

\item{\code{timeout}}{Default timeout in seconds for \pkg{chromote} to wait for a response.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{course}}{Returns or sets the course}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MoodlePage-is_valid_course_id}{\code{MoodlePage$is_valid_course_id()}}
\item \href{#method-MoodlePage-new}{\code{MoodlePage$new()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="auto_events_enable_args"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-auto_events_enable_args'><code>chromote::ChromoteSession$auto_events_enable_args()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="check_active"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-check_active'><code>chromote::ChromoteSession$check_active()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="close"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-close'><code>chromote::ChromoteSession$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="debug_log"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-debug_log'><code>chromote::ChromoteSession$debug_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_auto_events"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_auto_events'><code>chromote::ChromoteSession$get_auto_events()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_child_loop"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_child_loop'><code>chromote::ChromoteSession$get_child_loop()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_init_promise"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_init_promise'><code>chromote::ChromoteSession$get_init_promise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_session_id"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_session_id'><code>chromote::ChromoteSession$get_session_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_target_id"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_target_id'><code>chromote::ChromoteSession$get_target_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="get_viewport_size"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-get_viewport_size'><code>chromote::ChromoteSession$get_viewport_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="go_to"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-go_to'><code>chromote::ChromoteSession$go_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="invoke_event_callbacks"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-invoke_event_callbacks'><code>chromote::ChromoteSession$invoke_event_callbacks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="is_active"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-is_active'><code>chromote::ChromoteSession$is_active()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="mark_closed"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-mark_closed'><code>chromote::ChromoteSession$mark_closed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="new_session"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-new_session'><code>chromote::ChromoteSession$new_session()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="print"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-print'><code>chromote::ChromoteSession$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="respawn"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-respawn'><code>chromote::ChromoteSession$respawn()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="screenshot"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-screenshot'><code>chromote::ChromoteSession$screenshot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="screenshot_pdf"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-screenshot_pdf'><code>chromote::ChromoteSession$screenshot_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="send_command"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-send_command'><code>chromote::ChromoteSession$send_command()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="set_viewport_size"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-set_viewport_size'><code>chromote::ChromoteSession$set_viewport_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="view"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-view'><code>chromote::ChromoteSession$view()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chromote" data-topic="ChromoteSession" data-id="wait_for"><a href='../../chromote/html/ChromoteSession.html#method-ChromoteSession-wait_for'><code>chromote::ChromoteSession$wait_for()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoodlePage-is_valid_course_id"></a>}}
\if{latex}{\out{\hypertarget{method-MoodlePage-is_valid_course_id}{}}}
\subsection{Method \code{is_valid_course_id()}}{
Determine whether an object represents a valid Moodle internal course id number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoodlePage$is_valid_course_id(course_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{course_id}}{A scalar vector holding a positive integer value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MoodlePage-new"></a>}}
\if{latex}{\out{\hypertarget{method-MoodlePage-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MoodlePage} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoodlePage$new(site_url, timeout = 15, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_url}}{A scalar character vector holding the root URL for your Moodle instance}

\item{\code{timeout}}{Scalar numeric vector. Describes the number of second to wait for network events (e.g., for a web page to load).}

\item{\code{...}}{Arguments passed to \link[chromote:ChromoteSession]{chromote::ChromoteSession} constructor}
}
\if{html}{\out{</div>}}
}
}
}
