% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleR-package.R, R/moodler.R
\docType{package}
\name{moodleR-package}
\alias{moodleR}
\alias{moodleR-package}
\title{moodleR: Helper Functions to Work with 'Moodle' Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions to connect to a 'Moodle' database, cache relevant tables locally and generate learning analytics. 'Moodle' is an open source Learning Management System (LMS) developed by MoodleHQ. For more information about Moodle, visit \url{https://moodle.org}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions to connect to a 'Moodle' database, cache relevant tables locally and generate learning analytics. 'Moodle' is an open source Learning Management System (LMS) developed by MoodleHQ. For more information about Moodle, visit \url{https://moodle.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chi2labs/moodleR}
  \item \url{https://chi2labs.github.io/moodleR/}
}


Useful links:
\itemize{
  \item \url{https://github.com/chi2labs/moodleR}
  \item \url{https://chi2labs.github.io/moodleR/}
}

}
\author{
\strong{Maintainer}: Aleksander Dietrichson \email{dietrichson@gmail.com}

Other contributors:
\itemize{
  \item Chi Square Laboratories [copyright holder]
  \item Darko Miletic [contributor]
  \item Pablo Pagnone [contributor]
  \item Alex Ondrus [contributor]
}

}
\keyword{internal}
