% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml_helpers.R
\name{mxml_create_basic}
\alias{mxml_create_basic}
\title{Convert HTML output to the basic components of any question.}
\usage{
mxml_create_basic(node, type)
}
\arguments{
\item{node}{the section of html code corresponding to the question.}

\item{type}{character specifying the type of question to construct.}
}
\value{
an XML node to be inserted into a document.
}
\description{
HTML code generated from knitting together an RMarkdown document which
corresponds to all question types is converted to Moodle XML. This is
called from within \code{\link{html_to_moodle}}.
}
