% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml_helpers.R
\name{mxml_create_category}
\alias{mxml_create_category}
\title{Create Moodle XML code for category.}
\usage{
mxml_create_category(text)
}
\arguments{
\item{text}{The name of the category to create.}
}
\value{
an XML node to be inserted into a document.
}
\description{
Given the text for the name of a category, create Moodle XML code. This is
called from within \code{\link{html_to_moodle}}.
}
