% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml_helpers.R
\name{mxml_obtain_options}
\alias{mxml_obtain_options}
\title{Retrieve Moodle options from HTML.}
\usage{
mxml_obtain_options(node)
}
\arguments{
\item{node}{HTML node that contains the options.}
}
\value{
a named character vector containing the options.
}
\description{
Given the text with the options for the question, retrieve them. This is
called from within \code{\link{html_to_moodle}}.
}
