% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml_helpers.R
\name{mxml_read_description}
\alias{mxml_read_description}
\title{Convert HTML output for Description Questions into Moodle XML.}
\usage{
mxml_read_description(node)
}
\arguments{
\item{node}{the section of html code corresponding to the description
question.}
}
\value{
an XML node to be inserted into a document.
}
\description{
HTML code generated from knitting together an RMarkdown document which
corresponds to a Description Question is converted to Moodle XML. This is
called from within \code{\link{html_to_moodle}}.
}
