% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml_helpers.R
\name{mxml_read_dragdrop}
\alias{mxml_read_dragdrop}
\title{Convert HTML output for Drag and Drop Matching Questions into Moodle XML.}
\usage{
mxml_read_dragdrop(node, opts)
}
\arguments{
\item{node}{the section of html code corresponding to the calculated
question.}

\item{opts}{named character vector of options associated with this question,
generally the result of a call to \code{\link{mxml_obtain_options}}.}
}
\value{
an XML node to be inserted into a document.
}
\description{
HTML code generated from knitting together an RMarkdown document which
corresponds to a Drag and Drop Matching Question is converted to Moodle XML.
This is called from within \code{\link{html_to_moodle}}.
}
