% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{open_moodle}
\alias{open_moodle}
\title{Open Moodle in Chrome/Chromium}
\usage{
open_moodle(
  site_url,
  graphical = TRUE,
  username = NULL,
  password = NULL,
  org = NULL,
  twoFA = list(),
  timeout = 10
)
}
\arguments{
\item{site_url}{A scalar character vector holding the root URL for your Moodle instance}

\item{graphical}{A scalar logical vector. If \code{TRUE} (the default), a graphical Chrome/Chromium window is opened for the user to log in to Moodle. If \code{FALSE}, a headless Chrome/Chromium instance is launched. The \code{username} and \code{password} arguments must be supplied if \code{graphical = FALSE}.}

\item{username}{A scalar character vector holding the username you log in to Moodle with}

\item{password}{A scalar character vector holding the password you log in to Moodle with}

\item{org}{A scalar character vector holding the name of the organization you belong to. Used when logging in to a Moodle page the has federated SSO.}

\item{twoFA}{A list of named character vectors. The name of each element should describe a 2FA provider and each element should describe a 2FA method. See the Details section for  more information}

\item{timeout}{Scalar numeric vector. Describes the number of second to wait for network events (e.g., for a web page to load).}
}
\value{
A \link{MoodlePage} object
}
\description{
This function launches a (possibly headless) Chrome/Chromium instance, and opens the dashboard
page at the given Moodle instance.
}
\details{
Currently moodleR only supports Duo as a two-factor authentication provider, and only supports
authentication via OTP codes, hardware security key, or phone call. Thus the \code{twoFA} argument
must be one of:
\itemize{
\item \code{list(duo = "otp")}
\item \code{list(duo = "key")}
\item \code{list(duo = "call")}
}
}
\examples{
\dontrun{
tab <- open_moodle(
  site_url = "https://moodle.smith.edu",
  graphical = FALSE,
  username = "whopper",
  password = "CorrectHorseBatteryStaple",
  org = "smith",
  twoFA = list(duo = "push")
)
tab$view() # opens a debugging window showing the "headless" tab
}

}
