% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.data.R
\name{paste.data}
\alias{paste.data}
\title{paste.data}
\usage{
paste.data(sep = "", header = TRUE, is.table = FALSE)
}
\description{
This function is used to read data from copy into R
}
\examples{
paste.data()
}
\keyword{moodle}
