% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_plot_functions.R
\name{plot.mdl_grades}
\alias{plot.mdl_grades}
\title{Plot Moodle Grades}
\usage{
\method{plot}{mdl_grades}(x, ...)
}
\arguments{
\item{x}{a lazy tbl reference with class mdl_forum_posts}

\item{...}{passed to ggplot}
}
\value{
A ggplot
}
\description{
Histogram (density) of normalized grades
}
