% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{populate_groups}
\alias{populate_groups}
\title{Add members to existing Moodle groups}
\usage{
populate_groups(tab, groups)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{groups}{A data frame containing the group names and emails of students in each group. The name of the column holding the emails must be \code{email_address}, and the name of the column holding the group names must be \code{group_name}.

The \code{group_name} column must hold entries corresponding to existing Moodle group names. The list of existing groups in a Moodle course can be retrieved using \code{\link[=get_group_ids]{get_group_ids()}}}
}
\value{
An (invisible) list of \link[httr:response]{httr::response} objects containing the response from Moodle's group/members.php endpoint for each added group
}
\description{
Populates Moodle groups with members based on correspondence between Moodle group name, and a
"tidy" data frame describing the group each student is in.
}
