% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_num.R
\name{pretty_num}
\alias{pretty_num}
\title{Format with Orders of Magnitude}
\usage{
pretty_num(numbers, style = c("default", "nopad", "6"))
}
\arguments{
\item{numbers}{numeric vector to be formatted}

\item{style}{formatting style}
}
\value{
formatted charecter vector
}
\description{
Formats 1234 as 1.23 K for easier reading. This is a temporary function, awaiting a pull-request and CRAN publication of https://github.com/r-lib/prettyunits
}
