% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_print_functions.R
\name{print.mdl_courses_summary}
\alias{print.mdl_courses_summary}
\title{Print Moodle Courses Summary}
\usage{
\method{print}{mdl_courses_summary}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Overrides generic print function
}
