% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_forum.R
\name{print.mdl_forum}
\alias{print.mdl_forum}
\title{Print moodle forum}
\usage{
\method{print}{mdl_forum}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"mdl_forum"}}

\item{...}{further arguments passed to methods}
}
\description{
Print mdl_forum object.
}
