% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_forum.R
\name{print.mdl_forum_data}
\alias{print.mdl_forum_data}
\title{Print moodle forum data}
\usage{
\method{print}{mdl_forum_data}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"mdl_forum_data"}}

\item{...}{further arguments passed to methods}
}
\description{
Print mdl_forum_data object.
}
