% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{print.mdl_quiz}
\alias{print.mdl_quiz}
\title{Print moodle quiz}
\usage{
\method{print}{mdl_quiz}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"mdl_quiz"}}

\item{...}{further arguments passed to methods}
}
\description{
Print mdl_quiz object.
}
