% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_quiz.R
\name{print.mdl_quiz_data}
\alias{print.mdl_quiz_data}
\title{Print moodle quiz data}
\usage{
\method{print}{mdl_quiz_data}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"mdl_quiz"}}

\item{...}{further arguments passed to methods}
}
\description{
Print mdl_quiz object.
}
