% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qamatrix.R
\name{qamatrix}
\alias{qamatrix}
\title{qamatrix}
\usage{
qamatrix(tbl)
}
\arguments{
\item{tbl}{a matrix}
}
\description{
This function takes a matrix and generates the html code for questions and answers in a
moodel quiz
}
\examples{
qamatrix(x)
}
\keyword{moodle}
