% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{quiz_items.mdl_quiz_data}
\alias{quiz_items.mdl_quiz_data}
\title{Extract item-level and distractor data}
\usage{
\method{quiz_items}{mdl_quiz_data}(x, marks = "categorical",
  question.type = NULL, fill = NA, mat = TRUE)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz_data"}}

\item{marks}{Char, defaults to "categorical"}

\item{question.type}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}

\item{fill}{What to use to fill missing values, defaults to \code{NA}}

\item{mat}{If \code{TRUE} (default), the result will be a \code{matrix} otherwise a \code{data.frame}}
}
\description{
Extract a matrix or data.frame of question marks
}
\details{
Extract item marks either as binary (correct/incorrect) or categorical data (for multiplechoice questions). If the test contains a mixture of binary and categorical items (such as truefalse and multichoice), than \code{marks = "binary"} will collapse multiple choice items into correct/incorrect (1/0). When using \code{marks = "categorical"}, mutliplechoice options will be preserved but a key needs to be obtained (see \code{\link{quiz_key}} for details). Weighted percentages can also be extracted by \code{marks = "weighted"} - in this case, each item's contribution will be weighted by its nominal weight, set in the quiz edit page in Moodle. Finally, you can obtain nominal weights as they appear on the edit quiz page, using \code{marks = "nominal"}.
}
