% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{quiz_key}
\alias{quiz_key}
\title{Extract key}
\usage{
quiz_key(x, question.type = NULL, complete = FALSE)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz_data"}}

\item{question.type}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}

\item{complete}{Should question name, text etc. be provided? Returns a \code{data.frame}}
}
\description{
Extract key (only when distractor data are available).
}
