% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{quiz_scores}
\alias{quiz_scores}
\title{Extract whole-quiz scores}
\usage{
quiz_scores(x, marks = "binary")
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz_data"}}

\item{marks}{How items should be graded; see \code{\link{quiz_items}} for details}
}
\description{
Extract whole-quiz scores for each attempt.
}
\details{
See \code{\link{quiz_items}} on how the item marks can be specified.
}
