% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{render_progress}
\alias{render_progress}
\title{Report on compilation in a directory}
\usage{
render_progress(path = ".", ...)
}
\arguments{
\item{path}{path to assignment directory}

\item{...}{currently ignored}
}
\description{
Report on compilation in a directory
}
\examples{
dir <- "~/Dropbox/sds192-s20/SDS192- 01_202003-Homework 2 ggplot2-530786"
render_progress(dir)
}
