% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_summary_functions.R
\name{summary.mdl_courses}
\alias{summary.mdl_courses}
\title{Summary of mdl_courses Object}
\usage{
summary.mdl_courses(object, ...)
}
\arguments{
\item{object}{a lazy tbl reference with class mdl_grades}

\item{...}{currently ignored}
}
\value{
a tibble with the summary.
}
\description{
Provides summary statistics for moodle courses
}
