% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_summary_functions.R
\name{summary.mdl_grades}
\alias{summary.mdl_grades}
\title{Summary of mdl_grads Object}
\usage{
\method{summary}{mdl_grades}(object, ...)
}
\arguments{
\item{object}{a lazy tbl reference with class mdl_grades}

\item{...}{currently ignored}
}
\value{
a tibble with the summary.
}
\description{
Provides summary statistics for moodle grades
}
