% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_summary_functions.R
\name{summary.mdl_users}
\alias{summary.mdl_users}
\title{Summary of mdl_users Object}
\usage{
\method{summary}{mdl_users}(object, ...)
}
\arguments{
\item{object}{a lazy tbl reference with class mdl_grades}

\item{...}{currently ignored}
}
\value{
a tibble with the summary.
}
\description{
Provides summary statistics for moodle courses
}
