% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_file}
\alias{upload_file}
\title{Upload file to Moodle}
\usage{
upload_file(tab, section, title, path, visible = TRUE)
}
\arguments{
\item{tab}{A \code{\link[=MoodlePage]{MoodlePage()}} object obtained via the \code{\link[=open_moodle]{open_moodle()}} function}

\item{section}{Scalar character vector giving the name of the section where the file should be added}

\item{title}{Scalar character vector giving the title displayed for the file on the main page}

\item{path}{Scalar character vector giving the path to the file to be uploaded}

\item{visible}{Scalar logical vector indicating whether the file should be visible or hidden upon upload}
}
\description{
Upload file to Moodle
}
