% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcards_mxml.R
\name{wildcards_mxml}
\alias{wildcards_mxml}
\title{Create Moodle XML code for storing a dataset for a calculated question.}
\usage{
wildcards_mxml(data, shared = TRUE)
}
\arguments{
\item{data}{a data.frame storing the wildcard information. The variables
in the data.frame correspond to wildcards. Each row corresponds to a
unique set of linked values.}

\item{shared}{a vector of boolean values of the same length as the number of
columns in \code{data}. If an element is TRUE, the corresponding wildcard
values are shared among questions within the same category. If FALSE, they
are specific to this question. Per Moodle documentation, if questions within
the same category make use of shared datasets and have the same number of
datasets, the questions will be synchronized. Values are replicated as
necessary (default = TRUE).}
}
\value{
a character string containing XML code.
}
\description{
Calculated questions in Moodle store the wildcards in a dataset.
\code{wildcard_mxml} creates the Moodle XML node for storing the wildcards.
The code is returned as a text string that would generally be included in the
HTML file rendered by the RMarkdown process; it is then read and converted
to XML code using the \code{\link{mxml_read_calculated}} function.
}
