% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_xml_helpers.R
\name{word_create_calculated}
\alias{word_create_calculated}
\title{Convert Moodle XML of calculated questions into word format.}
\usage{
word_create_calculated(node, qnumber, solution, seed)
}
\arguments{
\item{node}{the section of html code corresponding to the question.}

\item{qnumber}{the question number to include.}

\item{solution}{a boolean indicating whether the solution/feedback should be
printed.}

\item{seed}{a numeric value to set the seed for determining the wildcard
values to use.}
}
\value{
a string of markdown text.
}
\description{
Moodle XML code that creates a question is converted to markdown syntax for
writing to Microsoft Word.
}
