% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_xml_helpers.R
\name{word_create_description}
\alias{word_create_description}
\title{Convert Moodle XML of description questions into word format.}
\usage{
word_create_description(node)
}
\arguments{
\item{node}{the section of html code corresponding to the question.}
}
\value{
a string of markdown text.
}
\description{
Moodle XML code that creates a question is converted to markdown syntax for
writing to Microsoft Word.
}
