% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_xml_helpers.R
\name{word_create_essay}
\alias{word_create_essay}
\title{Convert Moodle XML of essay questions into word format.}
\usage{
word_create_essay(node, qnumber, solution)
}
\arguments{
\item{node}{the section of html code corresponding to the question.}

\item{qnumber}{the question number to include.}

\item{solution}{a boolean indicating whether the solution/feedback should be
printed.}
}
\value{
a string of markdown text.
}
\description{
Moodle XML code that creates a question is converted to markdown syntax for
writing to Microsoft Word.
}
