% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{extract_key.mdl_quiz_data}
\alias{extract_key.mdl_quiz_data}
\title{Extract key}
\usage{
\method{extract_key}{mdl_quiz_data}(x, distractors = NULL)
}
\arguments{
\item{x}{Object of class \code{"mdl_quiz_data"}}

\item{distractors}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}
}
\description{
Extract key
}
