% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{feedback}
\alias{feedback}
\title{Generate feedback report for Moodle}
\usage{
feedback(x, ...)
}
\arguments{
\item{x}{key to Googlesheet}

\item{...}{arguments passed to \code{\link[googlesheets]{gs_key}}}
}
\description{
Generate feedback report for Moodle
}
\examples{
\dontrun{
key <- "1FA7XrFQ1h6Lss-c8__RvQceG0-vIO2vXR1hPcdD49-o"
feedback(key, visibility = "private")
}
}
