% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{feedback}
\alias{feedback}
\alias{add_feedback}
\alias{read_moodle_grades}
\alias{read_moodle_sheet}
\alias{join_moodle}
\title{Generate feedback report for Moodle}
\usage{
add_feedback(x, ...)

read_moodle_grades(...)

read_moodle_sheet(...)

join_moodle(x, y, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{arguments passed to other functions}

\item{y}{other data.frame}
}
\description{
Generate feedback report for Moodle
}
\examples{
\dontrun{
key <- "1S0vfV_xkebyTkC2bmhOv-MDNMLulb95GFwfvI9ZPtGw"
if (require(googlesheets4)) {
  # sheets_auth()
  grades <- key \%>\%
    read_sheet()
}
print_feedback(grades)
}
}
