% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_quiz.R
\name{get_module_data.mdl_quiz}
\alias{get_module_data.mdl_quiz}
\title{Get Quiz Data}
\usage{
\method{get_module_data}{mdl_quiz}(x, attempt = "first",
  question.type = NULL, distractors = TRUE, prefix = "mdl_",
  suppress.warnings = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"mdl_quiz"}}

\item{attempt}{Defaults to \code{"first"}; \code{"all"} and \code{"last"} are also allowed, or a numeric vector specifying attempt IDs}

\item{question.type}{Char vector, if \code{NULL}, all items regardless of type will be fetched}

\item{distractors}{List of distractors for to get key for; if \code{NULL} (the default) all keys will be extracted}

\item{prefix}{Defaults to \code{"mdl_"}}

\item{suppress.warnings}{Should warnings produced by \code{\link[DBI]{dbGetQuery}} be suppressed? Defaults to \code{TRUE}}

\item{...}{Further arguments passed on to methods}
}
\description{
Fetch item data from a quiz.
}
