% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{get_question_ans}
\alias{get_question_ans}
\title{Get Question Answer}
\usage{
get_question_ans(conn, question.type, attempt.id, prefix = "mdl_")
}
\arguments{
\item{conn}{Connection object}

\item{question.type}{Char vector}

\item{attempt.id}{Numeric vector}

\item{prefix}{Defaults to \code{"mdl_quiz"}}
}
\description{
Get question answer.
}
