% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_courses.R
\name{mdl_courses}
\alias{mdl_courses}
\title{Get Courses}
\usage{
mdl_courses(con = mdl_get_connection(), tbl_prefix = "mdl_")
}
\arguments{
\item{con}{a database connection object}

\item{tbl_prefix}{prefix for tables}
}
\value{
A dbplyr reference object.
}
\description{
Returns a reference to the (cached) course table, with the most relevant columns selected.
}
\details{
For convenience a join with the category table is made, and "category_name" added
}
