% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_grades.R
\name{mdl_grades}
\alias{mdl_grades}
\title{Get Grades}
\usage{
mdl_grades(con = mdl_get_connection())
}
\arguments{
\item{con}{a database connection object}
}
\value{
A dbplyr reference object
}
\description{
Returns a reference to the (cached) grades table, with the most relevant columns selected.
}
