% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{MoodleR-package}
\alias{MoodleR}
\alias{MoodleR-package}
\title{Construct Moodle quizzes from RMarkdown documents.}
\description{
Author quiz questions using RMarkdown and convert them into Moodle XLM code
for import into Moodle. This is particularly helpful if you want to assign
each student the same problem but with randomly generated data.
}
\details{
The primary function from \pkg{MoodleR} is \code{\link{html_to_moodle}}.
}
\author{
\strong{Maintainer}: Reyes Eric \email{reyesem@rose-hulman.edu}

}
