% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_plot_functions.R
\name{plot.mdl_forum_posts}
\alias{plot.mdl_forum_posts}
\title{Wordcloud Plot}
\usage{
\method{plot}{mdl_forum_posts}(x, ...)
}
\arguments{
\item{x}{a lazy tbl reference with class mdl_forum_posts}

\item{...}{passed to ggplot}
}
\value{
ggplot
}
\description{
Create a simple wordcloud plot based on a mdl_posts object.
}
