% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot_functions.R
\name{plot.mdl_forum_posts}
\alias{plot.mdl_forum_posts}
\title{Wordcloud Plot}
\usage{
\method{plot}{mdl_forum_posts}(x, use_base_graphics = FALSE, ...)
}
\arguments{
\item{x}{a lazy tbl reference or data.frame.}

\item{use_base_graphics}{Should base graphics be produces instead of ggplot?}

\item{...}{passed to ggplot or base graphics}
}
\value{
A ggplot
}
\description{
Create a simple wordcloud plot based on a mdl_posts object.
}
