% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot_functions.R
\name{plot.mdl_grades}
\alias{plot.mdl_grades}
\title{Plot Moodle Grades}
\usage{
\method{plot}{mdl_grades}(x, use_base_graphics = FALSE, ...)
}
\arguments{
\item{x}{a lazy tbl reference}

\item{use_base_graphics}{Should base graphics be used (instead of ggplot)}

\item{...}{passed along to ggplot or base graphic function}
}
\value{
A ggplot or base  graphic plot
}
\description{
Histogram (density) of normalized grades
}
