% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_summary_functions.R
\name{summary.mdl_forum_posts}
\alias{summary.mdl_forum_posts}
\title{Summary of mdl_forum_posts Object}
\usage{
\method{summary}{mdl_forum_posts}(object, ...)
}
\arguments{
\item{object}{a lazy tbl reference with class mdl_grades}

\item{...}{currently ignored}
}
\value{
a tibble with the summary.
}
\description{
Provides summary statistics for forum posts.
}
